v0.1.47 (alpha-discovery)
Bug discovered: When the customer is directly in line with the target on the X axis, The navmesh autobreak causes the customer to break so much that it would take 20 seconds for the player to reach the target.

To fix it, the customer X spawn position has been ofsetted by 0.01 give or take. It should never be 0.